+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:ebidence/firebase_options.dart';
 import 'package:ebidence/function/gif_recorder.dart';
 import 'package:ebidence/function/x_share.dart';
-import 'package:ebidence/view/developer/cloud_storage.dart';
-import 'package:ebidence/view/developer/send_db.dart';
+import 'package:ebidence/view/developer/send_firebase.dart';
+
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 
@@ -41,7 +41,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home: SendDb(),
+      home: SendFirebase(),
     );
   }
 }
